+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
---
+- Access Container from Kingdom instance
+- Added Kingdom Service support
+- Removed Veneer dependency
+- Simplified Cache, Config and Store dependency injection
 
 ---